refactor: modularize testing framework
