+++ src/core/testing_framework/testing_core.py
@@ -1,513 +1,61 @@
 #!/usr/bin/env python3
-"""
-Testing Core - V2 Testing Framework Core Classes
-================================================
+"""High-level orchestration for the testing framework."""
 
-Contains base classes and core test implementations for the testing framework.
-Follows V2 coding standards with clean OOP design and SRP compliance.
-
-Author: Agent-2 (Architecture & Design Specialist)
-License: MIT
-"""
+from __future__ import annotations
 
 import asyncio
-import time
 import logging
+from typing import Iterable, List
 
-from src.utils.stability_improvements import stability_manager, safe_import
-from abc import ABC, abstractmethod
-from typing import Dict, Any, List, Optional, Union
-from unittest.mock import Mock, AsyncMock
-
-# Import our types
-try:
-    from .testing_types import TestStatus, TestType, TestPriority, TestResult, TestScenario
-except ImportError:
-    from testing_types import TestStatus, TestType, TestPriority, TestResult, TestScenario
+from .base_test import BaseIntegrationTest
+from .integration_tests import (
+    CrossSystemCommunicationTest,
+    ServiceIntegrationTest,
+    DatabaseIntegrationTest,
+)
 
 
-class BaseIntegrationTest(ABC):
-    """Abstract base class for all integration tests."""
-    
-    def __init__(self, test_name: str, test_type: TestType, priority: TestPriority = TestPriority.NORMAL):
-        self.test_name = test_name
-        self.test_type = test_type
-        self.priority = priority
-        self.status = TestStatus.PENDING
-        self.start_time: Optional[float] = None
-        self.end_time: Optional[float] = None
-        self.error_message: Optional[str] = None
-        self.error_traceback: Optional[str] = None
-        self.metrics: Dict[str, Any] = {}
-        self.logs: List[str] = []
-        self.assertions_passed = 0
-        self.assertions_failed = 0
-        self.test_data: Dict[str, Any] = {}
-        
-        # Setup logging
-        self.logger = logging.getLogger(f"test.{self.__class__.__name__}")
-        self.logger.setLevel(logging.INFO)
-    
-    @abstractmethod
-    async def setup(self) -> bool:
-        """Setup test environment and dependencies."""
-        pass
-    
-    @abstractmethod
-    async def execute(self) -> bool:
-        """Execute the actual test logic."""
-        pass
-    
-    @abstractmethod
-    async def cleanup(self) -> bool:
-        """Cleanup test environment and resources."""
-        pass
-    
-    @abstractmethod
-    async def validate(self) -> bool:
-        """Validate test results and assertions."""
-        pass
-    
-    async def run(self) -> TestResult:
-        """Run the complete test lifecycle."""
-        self.status = TestStatus.RUNNING
-        self.start_time = time.time()
-        
-        try:
-            self.logger.info(f"Starting test: {self.test_name}")
-            self.logs.append(f"Test started at {time.strftime('%H:%M:%S')}")
-            
-            # Setup phase
-            if not await self.setup():
-                raise RuntimeError("Test setup failed")
-            
-            # Execute phase
-            if not await self.execute():
-                raise RuntimeError("Test execution failed")
-            
-            # Validate phase
-            if not await self.validate():
-                raise RuntimeError("Test validation failed")
-            
-            # Success
-            self.status = TestStatus.PASSED
-            self.logs.append("Test completed successfully")
-            
-        except Exception as e:
-            self.status = TestStatus.FAILED
-            self.error_message = str(e)
-            self.error_traceback = str(e.__traceback__)
-            self.logs.append(f"Test failed: {e}")
-            self.logger.error(f"Test failed: {e}")
-            
-        finally:
-            # Cleanup phase
-            try:
-                await self.cleanup()
-            except Exception as cleanup_error:
-                self.logs.append(f"Cleanup failed: {cleanup_error}")
-                self.logger.warning(f"Cleanup failed: {cleanup_error}")
-            
-            self.end_time = time.time()
-            duration = self.end_time - self.start_time
-            
-            self.logs.append(f"Test finished at {time.strftime('%H:%M:%S')}")
-            self.logger.info(f"Test {self.test_name} finished with status: {self.status.value}")
-        
-        return TestResult(
-            test_id=f"{self.__class__.__name__}_{int(time.time())}",
-            test_name=self.test_name,
-            test_type=self.test_type,
-            status=self.status,
-            start_time=self.start_time or 0,
-            end_time=self.end_time or 0,
-            duration=duration if self.start_time and self.end_time else 0,
-            error_message=self.error_message,
-            error_traceback=self.error_traceback,
-            metrics=self.metrics,
-            logs=self.logs,
-            assertions_passed=self.assertions_passed,
-            assertions_failed=self.assertions_failed,
-            test_data=self.test_data
-        )
+class TestManager:
+    """Lightweight orchestrator that runs integration tests sequentially."""
+
+    def __init__(self, tests: Iterable[BaseIntegrationTest] | None = None):
+        self.tests: List[BaseIntegrationTest] = list(tests or [])
+        self.logger = logging.getLogger(__name__)
+
+    def add_test(self, test: BaseIntegrationTest) -> None:
+        self.tests.append(test)
+        self.logger.info("Added test %s", test.test_name)
+
+    async def run_all(self):
+        results = []
+        for test in self.tests:
+            self.logger.info("Running %s", test.test_name)
+            results.append(await test.run())
+        return results
 
 
-class CrossSystemCommunicationTest(BaseIntegrationTest):
-    """Test for cross-system communication functionality."""
-    
-    def __init__(self, test_name: str, priority: TestPriority = TestPriority.HIGH):
-        super().__init__(test_name, TestType.INTEGRATION, priority)
-        self.communication_systems = []
-        self.message_queue = []
-        self.connection_status = {}
-    
-    async def setup(self) -> bool:
-        """Setup communication test environment."""
-        try:
-            self.logs.append("Setting up communication test environment")
-            
-            # Mock communication systems
-            self.communication_systems = [
-                {"name": "system_a", "status": "active"},
-                {"name": "system_b", "status": "active"},
-                {"name": "system_c", "status": "active"}
-            ]
-            
-            # Initialize message queue
-            self.message_queue = []
-            self.connection_status = {sys["name"]: "connected" for sys in self.communication_systems}
-            
-            self.logs.append(f"Setup {len(self.communication_systems)} communication systems")
-            return True
-            
-        except Exception as e:
-            self.logs.append(f"Setup failed: {e}")
-            return False
-    
-    async def execute(self) -> bool:
-        """Execute communication test."""
-        try:
-            self.logs.append("Executing communication test")
-            
-            # Test message sending
-            test_message = {"type": "test", "content": "Hello from system A", "timestamp": time.time()}
-            self.message_queue.append(test_message)
-            
-            # Simulate message processing
-            await asyncio.sleep(0.1)
-            
-            # Test message reception
-            if self.message_queue:
-                received_message = self.message_queue.pop(0)
-                if received_message["content"] == "Hello from system A":
-                    self.assertions_passed += 1
-                    self.logs.append("Message sending/receiving test passed")
-                else:
-                    self.assertions_failed += 1
-                    self.logs.append("Message sending/receiving test failed")
-            
-            # Test connection status
-            for system in self.communication_systems:
-                if self.connection_status.get(system["name"]) == "connected":
-                    self.assertions_passed += 1
-                else:
-                    self.assertions_failed += 1
-            
-            self.logs.append(f"Communication test completed: {self.assertions_passed} passed, {self.assertions_failed} failed")
-            return True
-            
-        except Exception as e:
-            self.logs.append(f"Execution failed: {e}")
-            return False
-    
-    async def cleanup(self) -> bool:
-        """Cleanup communication test environment."""
-        try:
-            self.logs.append("Cleaning up communication test environment")
-            
-            # Clear message queue
-            self.message_queue.clear()
-            
-            # Reset connection status
-            self.connection_status.clear()
-            
-            # Clear communication systems
-            self.communication_systems.clear()
-            
-            self.logs.append("Cleanup completed successfully")
-            return True
-            
-        except Exception as e:
-            self.logs.append(f"Cleanup failed: {e}")
-            return False
-    
-    async def validate(self) -> bool:
-        """Validate communication test results."""
-        try:
-            self.logs.append("Validating communication test results")
-            
-            # Check if all assertions passed
-            total_assertions = self.assertions_passed + self.assertions_failed
-            if total_assertions == 0:
-                self.logs.append("No assertions were made during test")
-                return False
-            
-            success_rate = self.assertions_passed / total_assertions
-            self.metrics["success_rate"] = success_rate
-            self.metrics["total_assertions"] = total_assertions
-            
-            self.logs.append(f"Validation completed: {success_rate:.2%} success rate")
-            return success_rate >= 0.8  # 80% success rate threshold
-            
-        except Exception as e:
-            self.logs.append(f"Validation failed: {e}")
-            return False
-
-
-class ServiceIntegrationTest(BaseIntegrationTest):
-    """Test for service integration functionality."""
-    
-    def __init__(self, test_name: str, priority: TestPriority = TestPriority.NORMAL):
-        super().__init__(test_name, TestType.INTEGRATION, priority)
-        self.services = []
-        self.service_status = {}
-        self.api_calls = []
-    
-    async def setup(self) -> bool:
-        """Setup service integration test environment."""
-        try:
-            self.logs.append("Setting up service integration test environment")
-            
-            # Mock services
-            self.services = [
-                {"name": "auth_service", "endpoint": "/auth", "status": "active"},
-                {"name": "user_service", "endpoint": "/users", "status": "active"},
-                {"name": "data_service", "endpoint": "/data", "status": "active"}
-            ]
-            
-            # Initialize service status
-            self.service_status = {svc["name"]: "healthy" for svc in self.services}
-            self.api_calls = []
-            
-            self.logs.append(f"Setup {len(self.services)} services")
-            return True
-            
-        except Exception as e:
-            self.logs.append(f"Setup failed: {e}")
-            return False
-    
-    async def execute(self) -> bool:
-        """Execute service integration test."""
-        try:
-            self.logs.append("Executing service integration test")
-            
-            # Test service health checks
-            for service in self.services:
-                if self.service_status.get(service["name"]) == "healthy":
-                    self.assertions_passed += 1
-                    self.logs.append(f"Service {service['name']} health check passed")
-                else:
-                    self.assertions_failed += 1
-                    self.logs.append(f"Service {service['name']} health check failed")
-            
-            # Test API calls
-            test_api_call = {
-                "service": "auth_service",
-                "endpoint": "/auth/login",
-                "method": "POST",
-                "status": 200
-            }
-            self.api_calls.append(test_api_call)
-            
-            if test_api_call["status"] == 200:
-                self.assertions_passed += 1
-                self.logs.append("API call test passed")
-            else:
-                self.assertions_failed += 1
-                self.logs.append("API call test failed")
-            
-            self.logs.append(f"Service integration test completed: {self.assertions_passed} passed, {self.assertions_failed} failed")
-            return True
-            
-        except Exception as e:
-            self.logs.append(f"Execution failed: {e}")
-            return False
-    
-    async def cleanup(self) -> bool:
-        """Cleanup service integration test environment."""
-        try:
-            self.logs.append("Cleaning up service integration test environment")
-            
-            # Clear services
-            self.services.clear()
-            self.service_status.clear()
-            self.api_calls.clear()
-            
-            self.logs.append("Cleanup completed successfully")
-            return True
-            
-        except Exception as e:
-            self.logs.append(f"Cleanup failed: {e}")
-            return False
-    
-    async def validate(self) -> bool:
-        """Validate service integration test results."""
-        try:
-            self.logs.append("Validating service integration test results")
-            
-            # Check if all assertions passed
-            total_assertions = self.assertions_passed + self.assertions_failed
-            if total_assertions == 0:
-                self.logs.append("No assertions were made during test")
-                return False
-            
-            success_rate = self.assertions_passed / total_assertions
-            self.metrics["success_rate"] = success_rate
-            self.metrics["total_assertions"] = total_assertions
-            self.metrics["services_tested"] = len(self.services)
-            
-            self.logs.append(f"Validation completed: {success_rate:.2%} success rate")
-            return success_rate >= 0.8  # 80% success rate threshold
-            
-        except Exception as e:
-            self.logs.append(f"Validation failed: {e}")
-            return False
-
-
-class DatabaseIntegrationTest(BaseIntegrationTest):
-    """Test for database integration functionality."""
-    
-    def __init__(self, test_name: str, priority: TestPriority = TestPriority.NORMAL):
-        super().__init__(test_name, TestType.INTEGRATION, priority)
-        self.database_connections = []
-        self.test_queries = []
-        self.query_results = {}
-    
-    async def setup(self) -> bool:
-        """Setup database integration test environment."""
-        try:
-            self.logs.append("Setting up database integration test environment")
-            
-            # Mock database connections
-            self.database_connections = [
-                {"name": "main_db", "type": "postgresql", "status": "connected"},
-                {"name": "cache_db", "type": "redis", "status": "connected"},
-                {"name": "analytics_db", "type": "mongodb", "status": "connected"}
-            ]
-            
-            # Initialize test queries
-            self.test_queries = [
-                "SELECT COUNT(*) FROM users",
-                "SELECT * FROM users LIMIT 1",
-                "INSERT INTO test_table (name) VALUES ('test')"
-            ]
-            
-            self.query_results = {}
-            
-            self.logs.append(f"Setup {len(self.database_connections)} database connections")
-            return True
-            
-        except Exception as e:
-            self.logs.append(f"Setup failed: {e}")
-            return False
-    
-    async def execute(self) -> bool:
-        """Execute database integration test."""
-        try:
-            self.logs.append("Executing database integration test")
-            
-            # Test database connections
-            for db in self.database_connections:
-                if db["status"] == "connected":
-                    self.assertions_passed += 1
-                    self.logs.append(f"Database {db['name']} connection test passed")
-                else:
-                    self.assertions_failed += 1
-                    self.logs.append(f"Database {db['name']} connection test failed")
-            
-            # Test query execution
-            for query in self.test_queries:
-                # Simulate query execution
-                if "SELECT" in query:
-                    self.query_results[query] = {"status": "success", "rows": 1}
-                    self.assertions_passed += 1
-                    self.logs.append(f"Query '{query[:30]}...' executed successfully")
-                elif "INSERT" in query:
-                    self.query_results[query] = {"status": "success", "affected_rows": 1}
-                    self.assertions_passed += 1
-                    self.logs.append(f"Query '{query[:30]}...' executed successfully")
-                else:
-                    self.assertions_failed += 1
-                    self.logs.append(f"Query '{query[:30]}...' failed")
-            
-            self.logs.append(f"Database integration test completed: {self.assertions_passed} passed, {self.assertions_failed} failed")
-            return True
-            
-        except Exception as e:
-            self.logs.append(f"Execution failed: {e}")
-            return False
-    
-    async def cleanup(self) -> bool:
-        """Cleanup database integration test environment."""
-        try:
-            self.logs.append("Cleaning up database integration test environment")
-            
-            # Clear database connections
-            self.database_connections.clear()
-            self.test_queries.clear()
-            self.query_results.clear()
-            
-            self.logs.append("Cleanup completed successfully")
-            return True
-            
-        except Exception as e:
-            self.logs.append(f"Cleanup failed: {e}")
-            return False
-    
-    async def validate(self) -> bool:
-        """Validate database integration test results."""
-        try:
-            self.logs.append("Validating database integration test results")
-            
-            # Check if all assertions passed
-            total_assertions = self.assertions_passed + self.assertions_failed
-            if total_assertions == 0:
-                self.logs.append("No assertions were made during test")
-                return False
-            
-            success_rate = self.assertions_passed / total_assertions
-            self.metrics["success_rate"] = success_rate
-            self.metrics["total_assertions"] = total_assertions
-            self.metrics["databases_tested"] = len(self.database_connections)
-            self.metrics["queries_executed"] = len(self.test_queries)
-            
-            self.logs.append(f"Validation completed: {success_rate:.2%} success rate")
-            return success_rate >= 0.8  # 80% success rate threshold
-            
-        except Exception as e:
-            self.logs.append(f"Validation failed: {e}")
-            return False
+async def run_demo_suite():
+    """Run a small demonstration suite of built-in tests."""
+    manager = TestManager(
+        [
+            CrossSystemCommunicationTest("Communication Test"),
+            ServiceIntegrationTest("Service Test"),
+            DatabaseIntegrationTest("Database Test"),
+        ]
+    )
+    return await manager.run_all()
 
 
 def run_smoke_test() -> bool:
-    """Run smoke test for testing core module."""
+    """Execute the demo suite to verify orchestration works."""
     try:
-        print("🧪 Testing Testing Core Module...")
-        
-        # Test base class instantiation (should fail for abstract class)
-        try:
-            test = BaseIntegrationTest("test", TestType.INTEGRATION)
-            print("❌ BaseIntegrationTest should not be instantiable")
-            return False
-        except TypeError:
-            print("✅ BaseIntegrationTest correctly prevents instantiation")
-        
-        # Test CrossSystemCommunicationTest
-        comm_test = CrossSystemCommunicationTest("Communication Test")
-        assert comm_test.test_name == "Communication Test"
-        assert comm_test.test_type == TestType.INTEGRATION
-        assert comm_test.priority == TestPriority.HIGH
-        print("✅ CrossSystemCommunicationTest creation successful")
-        
-        # Test ServiceIntegrationTest
-        service_test = ServiceIntegrationTest("Service Test")
-        assert service_test.test_name == "Service Test"
-        assert service_test.test_type == TestType.INTEGRATION
-        print("✅ ServiceIntegrationTest creation successful")
-        
-        # Test DatabaseIntegrationTest
-        db_test = DatabaseIntegrationTest("Database Test")
-        assert db_test.test_name == "Database Test"
-        assert db_test.test_type == TestType.INTEGRATION
-        print("✅ DatabaseIntegrationTest creation successful")
-        
-        print("🎉 All testing core smoke tests passed!")
+        asyncio.run(run_demo_suite())
+        print("🎉 Testing core smoke test passed!")
         return True
-        
-    except Exception as e:
-        print(f"❌ Testing core smoke test failed: {e}")
+    except Exception as exc:  # pragma: no cover - defensive
+        print(f"❌ Testing core smoke test failed: {exc}")
         return False
 
 
-if __name__ == "__main__":
+if __name__ == "__main__":  # pragma: no cover
     run_smoke_test()
