+++ src/core/testing_framework/testing_orchestration_core.py
@@ -20,11 +20,25 @@ from src.utils.stability_improvements import stability_manager, safe_import
 
 # Import our types and core classes
 try:
-    from .testing_types import TestStatus, TestType, TestPriority, TestResult, TestScenario, TestEnvironment
-    from .testing_core import BaseIntegrationTest
+    from .testing_types import (
+        TestStatus,
+        TestType,
+        TestPriority,
+        TestResult,
+        TestScenario,
+        TestEnvironment,
+    )
+    from .base_test import BaseIntegrationTest
 except ImportError:
-    from testing_types import TestStatus, TestType, TestPriority, TestResult, TestScenario, TestEnvironment
-    from testing_core import BaseIntegrationTest
+    from testing_types import (
+        TestStatus,
+        TestType,
+        TestPriority,
+        TestResult,
+        TestScenario,
+        TestEnvironment,
+    )
+    from base_test import BaseIntegrationTest
 
 
 @dataclass
